update: prefinal stage

+++ src/wishes/wishes.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateWishDto } from './dto/create-wish.dto';
 import { UpdateWishDto } from './dto/update-wish.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Wish } from './entities/wish.entity';
-import { Repository } from 'typeorm';
+import { FindManyOptions, Repository } from 'typeorm';
+import { User } from '../users/entities/user.entity';
 
 @Injectable()
 export class WishesService {
@@ -12,8 +13,11 @@ export class WishesService {
     private wishRepository: Repository<Wish>,
   ) {}
 
-  async create(createWishDto: CreateWishDto) {
-    return this.wishRepository.save(createWishDto);
+  async create(user: User, createWishDto: CreateWishDto) {
+    return this.wishRepository.save({
+      ...createWishDto,
+      owner: user,
+    });
   }
 
   async getTop() {
@@ -28,18 +32,68 @@ export class WishesService {
     return this.wishRepository.findOneBy({ id });
   }
 
-  async updateById(id: number, updateWishDto: UpdateWishDto) {
-    return this.wishRepository.update({ id }, updateWishDto);
+  async findOne(query: FindManyOptions<Wish>) {
+    const wish = await this.wishRepository.findOne(query);
+    if (!wish) throw new NotFoundException('Wish not found');
+    return wish;
   }
 
-  async deleteById(id: number) {
-    return this.wishRepository.delete({ id });
+  async findMany(query: FindManyOptions<Wish>) {
+    const wishes = await this.wishRepository.find(query);
+
+    if (!wishes) throw new NotFoundException('Wish not found');
+    return wishes;
   }
 
-  async copy(id: number, updateWishDto: UpdateWishDto) {
-    return await this.wishRepository.update(
+  async updateById(user: User, id: number, updateWishDto: UpdateWishDto) {
+    try {
+      const wish = await this.wishRepository.findOne({ where: { id } });
+      if (wish.owner.id === user.id)
+        return this.wishRepository.update({ id }, updateWishDto);
+    } catch {
+      throw new NotFoundException('Wish not found');
+    }
+  }
+
+  async deleteById(user: User, id: number) {
+    try {
+      const wish = await this.wishRepository.findOne({ where: { id } });
+      if (wish.owner.id === user.id) return this.wishRepository.delete({ id });
+    } catch {
+      throw new NotFoundException('Wish not found');
+    }
+  }
+
+  async raise(id: number, offerSum: number) {
+    const wish = await this.wishRepository.findOne({ where: { id } });
+    if (!wish) throw new NotFoundException('Wish not found');
+
+    return this.wishRepository.update(
       { id },
-      { copied: () => 'copied + 1' },
+      { raised: wish.raised + offerSum },
     );
+  }
+
+  async copy(user, id: number) {
+    const wish = await this.wishRepository.findOne({ where: { id } });
+    if (!wish) throw new NotFoundException('Wish not found');
+
+    await this.wishRepository.update(
+      { id },
+      {
+        copied: wish.copied + 1,
+      },
+    );
+
+    return this.wishRepository.create({
+      name: wish.name,
+      link: wish.link,
+      image: wish.image,
+      price: wish.price,
+      description: wish.description,
+      owner: user,
+      copied: 0,
+      raised: 0,
+    });
   }
 }
